ME-52 Запрет перехвата объекта
Dev:

- Переработана система интерактивности. Вместо Animator теперь передается GameObject вызывающего интерактивность объекта
- Добавлен компонент проверяющий возможность взаимодействия с объектом `IInteractionAccess`. Создано две реализации. `DefaultInteractionAccess` проверяет, что в локальной сцене объект доступен к взаимодействию (боты не взаимодействуют с данным объектом). `MVNetworkInteractionAccess` добавляет к проверке взаимодействия в локальной сцене сетевой компонент (другие игроки в сети не взаимодействуют с данным объектом).
- В `NetworkObject` добавлен флаг CanBeOwned, отвечающий за добавление `MVNetworkInteractionAccess` к сетевому объекту. Данный флаг ставится тогда, когда нужно разрешить взаимодействие с объектом только одному игроку (когда есть длительная интерактивность - захват объекта)
- Модифицированы скрипты `GrabableTransform` и `GrabableWithPhysics`. В них добавлена логика запроса на взаимодействие и запрет на взаимодействие с объектом, если с ним взаимодействует другой интерактор
- Созданы префабы кубов с `GrabableTransform` и `GrabableWithPhysics`. У данных префабов в `NetworkObject` установлен флаг CanBeOwned.
