+++ Assets/ApplicationContent/APIScripts/External/Interactions/DefaultInteractionAccess.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 80c39ef928c54ebfb4da97d0a2ed5229
+timeCreated: 1746087613
