+++ Assets/ApplicationContent/APIScripts/External/Interactions/IInteractionAccess.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 1564bbdfb64e414691f3ec3b6b2733cc
+timeCreated: 1746086763
