+++ Assets/ApplicationContent/EmbeddedScenes/Adam/SceneObjects/Ball/Ball.prefab
@@ -14,10 +14,10 @@ GameObject:
   - component: {fileID: 5834091247010114144}
   - component: {fileID: 1781749843927196806}
   - component: {fileID: -4334892181520885375}
-  - component: {fileID: 3857218208953208150}
-  - component: {fileID: 8109429212788535113}
-  - component: {fileID: 5646734952812852901}
-  - component: {fileID: 4057526001232568257}
+  - component: {fileID: 322281642260943050}
+  - component: {fileID: 5946257313198955945}
+  - component: {fileID: 5997129750351548054}
+  - component: {fileID: 2418623586520632152}
   m_Layer: 0
   m_Name: Ball
   m_TagString: Untagged
@@ -135,7 +135,8 @@ MonoBehaviour:
   _animatedObject: 0
   _transformObject: 0
   _physicObject: 1
---- !u!114 &3857218208953208150
+  _canBeOwned: 0
+--- !u!114 &322281642260943050
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -147,7 +148,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d48dfe3bbbaf4ba9af5e5e40add28218, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &8109429212788535113
+--- !u!114 &5946257313198955945
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -164,7 +165,7 @@ MonoBehaviour:
   serverOnly: 0
   visibility: 0
   hasSpawned: 0
---- !u!114 &5646734952812852901
+--- !u!114 &5997129750351548054
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -180,7 +181,7 @@ MonoBehaviour:
   syncMode: 0
   syncInterval: 0
   _ownershipTime: 0.2
---- !u!114 &4057526001232568257
+--- !u!114 &2418623586520632152
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -197,5 +198,4 @@ MonoBehaviour:
   syncInterval: 0
   _positionThreshold: 0.001
   _rotationThreshold: 0.1
-  _ownershipDeviationThreshold: 0.003
   _target: {fileID: 1781749843927196806}
