+++ Assets/ApplicationContent/EmbeddedScenes/Adam/SceneObjects/Cube/GrabableCube.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2b1e5bdfa9f937d4eb730349c6ab205d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
