+++ Assets/ApplicationContent/EmbeddedScenes/Adam/SceneObjects/Cube/GrabablePhysicCube.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b73c62b613889ca4bb0753672eb9a98b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
