--- Assets/ApplicationContent/EmbeddedScenes/Adam/SceneObjects/Cube/Scripts/GrabableObjects.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 83786dee7fc3462aaf52a785e4260a28
-timeCreated: 1745055143
