+++ Assets/ApplicationContent/NetworkCore/MirrorNetworking/Synchronization/Animations.meta
